+++ otherlibs/stdune/ansi_color.ml
@@ -123,18 +123,28 @@ module Style = struct
     Printf.sprintf "\027[%sm" (String.concat l ~sep:";")
 end
 
-let term_supports_color =
-  lazy
-    (match Stdlib.Sys.getenv "TERM" with
+let supports_color fd =
+  let is_smart =
+    match Stdlib.Sys.getenv "TERM" with
     | exception Not_found -> false
     | "dumb" -> false
-    | _ -> true)
+    | _ -> true
+  and clicolor =
+    match Stdlib.Sys.getenv "CLICOLOR" with
+    | exception Not_found -> true
+    | "0" -> false
+    | _ -> true
+  and clicolor_force =
+    match Stdlib.Sys.getenv "CLICOLOR_FORCE" with
+    | exception Not_found -> false
+    | "0" -> false
+    | _ -> true
+  in
+  (is_smart && Unix.isatty fd && clicolor) || clicolor_force
 
-let stdout_supports_color =
-  lazy (Lazy.force term_supports_color && Unix.isatty Unix.stdout)
+let stdout_supports_color = lazy (supports_color Unix.stdout)
 
-let stderr_supports_color =
-  lazy (Lazy.force term_supports_color && Unix.isatty Unix.stderr)
+let stderr_supports_color = lazy (supports_color Unix.stderr)
 
 let rec tag_handler current_styles ppf styles pp =
   Format.pp_print_as ppf 0 (Style.escape_sequence_no_reset styles);
@@ -162,22 +172,24 @@ let prerr =
 let strip str =
   let len = String.length str in
   let buf = Buffer.create len in
-  let rec loop i =
-    if i = len then Buffer.contents buf
+  let rec loop start i =
+    if i = len then (
+      if i - start > 0 then Buffer.add_substring buf str start (i - start);
+      Buffer.contents buf)
     else
-      match str.[i] with
-      | '\027' -> skip (i + 1)
-      | c ->
-        Buffer.add_char buf c;
-        loop (i + 1)
+      match String.unsafe_get str i with
+      | '\027' ->
+        if i - start > 0 then Buffer.add_substring buf str start (i - start);
+        skip (i + 1)
+      | _ -> loop start (i + 1)
   and skip i =
     if i = len then Buffer.contents buf
     else
-      match str.[i] with
-      | 'm' -> loop (i + 1)
+      match String.unsafe_get str i with
+      | 'm' -> loop (i + 1) (i + 1)
       | _ -> skip (i + 1)
   in
-  loop 0
+  loop 0 0
 
 let index_from_any str start chars =
   let n = String.length str in