+++ test/blackbox-tests/test-cases/melange/private-module.t/run.t
@@ -1,2 +1,7 @@
-Private modules are not supported, so X is accessible and build does not fail
-  $ dune build
+X is not accessible as it is private
+  $ dune build app/.app.objs/melange/app__Foo.js
+  File "app/foo.ml", line 1, characters 0-9:
+  1 | Lib.X.run ();;
+      ^^^^^^^^^
+  Error: The module Lib.X is an alias for module Lib__X, which is missing
+  [1]