+++ test/blackbox-tests/test-cases/watching/fs-memo.t
@@ -32,7 +32,7 @@ when necessary.
   >   (glob_files dir/file-?)
   >   (glob_files dir/subdir/file-?))
   >  (target result)
-  >  (action (bash "\| echo Executing rule...
+  >  (action (system "\| echo Executing rule...
   >                "\| echo %{deps}       |
   >                "\|   tr ' ' '\n'      |
   >                "\|   xargs -n 1       |